% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_dbh.R
\name{av_dbh}
\alias{av_dbh}
\title{Calculate average dbh at the sample plot}
\usage{
av_dbh(plot_id, tree_id, species, age, layer, dbh, height, only_measured_h = T)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{tree_id}{Unique tree id at the plot}

\item{species}{Tree species}

\item{age}{Tree age (years)}

\item{layer}{Stand structure layer}

\item{dbh}{Tree DBH (cm)}

\item{height}{Tree height (m) (you have to provide tree heights if only_measured_h = T)}

\item{only_measured_h}{if TRUE only trees, with measured height will be used (logical, default TRUE)}
}
\value{
Numeric vector of average DBH values (cm)
}
\description{
Calculate arithmetic mean diameter at breast height (DBH) per species,
age and layer group within each plot.
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
tree_id = c(1, 2 ,3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
age = c(40, 40, 40, 40, 60, 45, 50, 50, 50, 60, 60, 60, 60, 60)
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
dbh = c(33, 32, 31, NA, 28, 47, 12, 22, 41, 48, 27, 42, 25, 33)
av_dbh(plot_id, tree_id, species, age, layer, dbh, height)
}
