% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_height.R
\name{av_height}
\alias{av_height}
\title{Calculate average height at the sample plot}
\usage{
av_height(plot_id, species, age, layer, height)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{species}{Tree species}

\item{age}{Tree age (years)}

\item{layer}{Stand structure layer}

\item{height}{Tree height (m)}
}
\value{
Numeric vector of average tree heights (m)
}
\description{
Calculate arithmetic mean tree height per species,
age and layer group within each plot.
Missing heights (NA) are excluded from the calculation.
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
age = c(40, 40, 40, 40, 60, 45, 50, 50, 50, 60, 60, 60, 60, 60)
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
av_height(plot_id, species, age, layer, height)
}
