% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed_time.R
\name{elapsed_time}
\alias{elapsed_time}
\title{Returns elapsed time}
\usage{
elapsed_time(t0)
}
\arguments{
\item{t0}{Output of \code{proc.time}, either all elements or just element "elapsed".
No default.}
}
\description{
Given a time stamp generated with \code{proc.time}, returns elapsed time as a \code{string}
in sec., min. or hrs, as appropriate.
}
\examples{
start <- proc.time()
elapsed_time(t0=start)

start <- proc.time()[3]
elapsed_time(t0=start)
}
