% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_growth.R
\name{h_growth}
\alias{h_growth}
\title{Height growth prediction}
\usage{
h_growth(T1, T2, H1, species)
}
\arguments{
\item{T1}{Vector of Age with known height (years)}

\item{T2}{Vector of age when we want predict height (years)}

\item{H1}{Vector of Height at age T1 - top or average (m)}

\item{species}{Vector of Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB',
'JD', 'SO', 'ŚW'}
}
\value{
Numeric vector of predicted heights at age T2 (m)
}
\description{
Predict tree height at a target age (T2) based on current age (T1),
current height (H1) and species. Can also be used to calculate Site Index (SI)
by setting T2 as base age (usually 50 or 100).
}
\examples{
T1 = 100
T2 = 102
H1 = 32
species = 'SO'

h_growth(T1, T2, H1, species)

# Calculation of Site Index (SI)
T1 = c(85, 110)
T2 = c(100, 100)
H1 = c(27, 36)
species = c('DB', 'ŚW')

si <- h_growth(T1, T2, H1, species)
print(si)
}
