% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olkusz_v2.R
\docType{data}
\name{olkusz_v2}
\alias{olkusz_v2}
\title{Example dataset: Stand-level data for height growth modelling}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{stand_id}{Unique ID of the plot}
\item{species}{Tree species in Polish abbreviations}
\item{age}{Stand age}
\item{height}{Tree height (m)}
\item{volume}{Stand volume (m³/ha)}
\item{region}{Forest region code}
}
}
\usage{
data(olkusz_v2)
}
\description{
Stand-level forest data from the Olkusz area used to demonstrate
height growth prediction and Site Index calculation with th_growth().
}
\keyword{datasets}
