% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_v.R
\name{predict_v}
\alias{predict_v}
\title{Height growth prediction}
\usage{
predict_v(T1, T2, H1, V, species, region)
}
\arguments{
\item{T1}{Age with known height}

\item{T2}{Age when we want predict height}

\item{H1}{Height at age T1 (average)}

\item{V}{Volume at age T1}

\item{species}{Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB'}
}
\value{
Volume at age T2
}
\description{
Predict tree volume in T2, based on the T1, H1, region, species
}
\examples{
T2 = c(100, 101, 102)
T1 = c(102, 103, 104)
H1 = c(32, 33, 34)
V = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
region = c('I', 'II', 'GLOB')

predict_v(T1, T2, H1, V, species, region)
}
