% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spg.R
\name{spg}
\alias{spg}
\title{Total coarse wood production}
\usage{
spg(age, height, species, region = "GLOB")
}
\arguments{
\item{age}{tree stand age (years)}

\item{height}{tree stand top height (m)}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB' (default, global parameters)}
}
\value{
Numeric vector of total coarse wood production (m3/ha)
}
\description{
Predict total coarse wood production in the stand based on
the stand age, height and species. You can calculate the production using
global or region-specific parameters (Polish natural forest regions I-VIII).
}
\examples{
age = c(100, 101, 102)
height = c(32, 33, 34)
species = c('SO', 'DB', 'BK')
spg(age, height, species)
}
