% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_growth_test.R
\docType{data}
\name{years}
\alias{years}
\title{Volume prediction}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
years
}
\arguments{
\item{stand_id}{stand id}

\item{years}{number of years that we want to predict the growth}

\item{age}{tree stand age}

\item{height}{Height at age T1 (average)}

\item{volume}{Volume at age T1}

\item{species}{Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB' (default)}

\item{output_type}{by default results will be returned as a vector, you can type 'df' to return data frame}
}
\value{
Growth sum in specific timeline
}
\description{
Predict tree volume growth in specific period of time based on H1, region, species
}
\examples{
years = 10
stand_id = c(1, 2, 3)
age = c(100, 101, 102)
height = c(32, 33, 34)
volume = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
region = c('I', 'II', 'GLOB')
v_growth(stand_id, years, age, height, volume, species, region, output_type = 'df')
}
\keyword{datasets}
