% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zd.R
\name{zd}
\alias{zd}
\title{Stocking index}
\usage{
zd(volume_stand, vt_stand)
}
\arguments{
\item{volume_stand}{calculated stand volume}

\item{vt_stand}{calculated tabular volume (see v_tab() function)}
}
\value{
Stock index as a ratio of actual stock to tabular volume
}
\description{
Calculation of the stocking index of the stand on the basis of
the calculated volume and the tabular volume (see v_tab() function).
}
\examples{
volume_stand = c(250, 341, 80)
vt_stand = c(261, 182, 34)
zd(volume_stand, vt_stand)
}
