% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zd_share.R
\name{zd_share}
\alias{zd_share}
\title{Stocking index}
\usage{
zd_share(stand_id, volume, age, height, species)
}
\arguments{
\item{stand_id}{stand id}

\item{volume}{stand volume by age-species group (m3/ha)}

\item{age}{tree stand age by age-species group (years)}

\item{height}{tree stand height by age-species group (m)}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}
}
\value{
Numeric vector of stocking index values (dimensionless ratio of actual stock to tabular volume)
}
\description{
Calculate the stocking index (relative density) of the stand
as the ratio of actual stand volume to species-share-weighted tabular volume
(see v_tab()). Values above 1.0 indicate overstocking, below 1.0 understocking.
}
\examples{
stand_id = c(1, 2, 3)
age = c(100, 101, 102)
height = c(32, 33, 34)
volume = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
}
