% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_dbh.R
\name{av_dbh}
\alias{av_dbh}
\title{Calculate average dbh at the sample plot}
\usage{
av_dbh(plot_id, species, layer, dbh)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{species}{Tree species}

\item{layer}{Stand structure layer}

\item{dbh}{Tree dbh}
}
\value{
Average tree height
}
\description{
Calculate average dbh of the species at the plot in the layer
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
dbh = c(33, 32, 31, NA, 28, 47, 12, 22, 41, 48, 27, 42, 25, 33)
av_dbh(plot_id, species, layer, dbh)
}
