% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_height.R
\name{av_height}
\alias{av_height}
\title{Calculate average height at the sample plot}
\usage{
av_height(plot_id, species, layer, height)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{species}{Tree species}

\item{layer}{Stand structure layer}

\item{height}{Tree height}
}
\value{
Average tree height
}
\description{
Calculate average height of the species at the plot in the layer
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
av_dbh(plot_id, species, layer, height)
}
