% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_growth.R
\name{h_growth}
\alias{h_growth}
\title{Height growth prediction}
\usage{
h_growth(T1, T2, H1, species)
}
\arguments{
\item{T1}{Vector of Age with known height}

\item{T2}{Vector of age when we want predict height}

\item{H1}{Vector of Height at age T1 (Top or average)}

\item{species}{Vector of Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}
}
\value{
H2 - Height at age T2
}
\description{
Predict tree height in T2, based on the T1, H1 and species
}
\examples{
T1 = 100
T2 = 102
H1 = 32
species = 'SO'

h_growth(T1, T2, H1, species)
}
