% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorey_height.R
\name{lorey_height}
\alias{lorey_height}
\title{Calculate Lorey's height at the sample plot}
\usage{
lorey_height(plot_id, species, age, layer, height, dbh)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{species}{Tree species}

\item{age}{Tree age}

\item{layer}{Stand structure layer}

\item{height}{Tree height}

\item{dbh}{Tree dbh}
}
\value{
Average tree height
}
\description{
Calculate dbh weighted average height of the species at the plot in the layer and age group
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
age = c(40, 40, 40, 40, 60, 45, 50, 50, 50, 60, 60, 60, 60, 60)
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
dbh = c(33, 32, 31, NA, 28, 47, 12, 22, 41, 48, 27, 42, 25, 33)
lorey_height(plot_id, species, age, layer, height, dbh)
}
