% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olkusz_v1.R
\docType{data}
\name{olkusz_v1}
\alias{olkusz_v1}
\title{Example dataset: Stand-level data for volume growth modelling}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{stand_id}{Unique ID of the plot}
\item{species}{Tree species in Polish abbreviations}
\item{age}{Stand age}
\item{height}{Tree height (m)}
\item{volume}{Stand volume (m³/ha)}
\item{region}{Forest region code}
}
}
\usage{
data(olkusz_v1)
}
\description{
Stand-level forest data from the Olkusz area used to demonstrate
volume growth prediction with v_growth() and related stand-level functions
}
\keyword{datasets}
