% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olkusz_v2.R
\docType{data}
\name{olkusz_v2}
\alias{olkusz_v2}
\title{Example dataset: Sample forest plots}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{stand_id}{Unique ID of the plot}
\item{species}{Tree species in Polish abbreviations}
\item{age}{Stand age}
\item{height}{Tree height (m)}
\item{volume}{Stand volume (m³/ha)}
\item{region}{Forest region code}
}
}
\usage{
data(olkusz_v2)
}
\description{
This dataset contains example plot-level data used to demonstrate
the functionality of the silvaR package.
}
\keyword{datasets}
