% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_growth.R
\name{predict_v}
\alias{predict_v}
\title{Volume prediction}
\usage{
predict_v(years, age, H1, V, species, region)
}
\arguments{
\item{years}{number of years that we want to predict the growth}

\item{age}{tree stand age}

\item{H1}{Height at age T1 (average)}

\item{V}{Volume at age T1}

\item{species}{Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB'}
}
\value{
Growth sum in specific timeline
}
\description{
Predict tree volume in T2, based on the T1, H1, region, species
}
\examples{
years = 10
age = c(100, 101, 102)
H1 = c(32, 33, 34)
V = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
region = c('I', 'II', 'GLOB')

predict_v(years, age, H1, V, species, region)
}
