% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_origin.R
\name{set_origin}
\alias{set_origin}
\title{Set lowland or mountainous origin}
\usage{
set_origin(value)
}
\arguments{
\item{value}{forest habitat or region}
}
\value{
Species group name
}
\description{
Set lowland or mountainous origin
Based on the forest habitat or forest region
}
\examples{
list = c('1', 'II', 'Karpacka', 'Bór górski wilgotny', '14', 'BMB', 'Bór')
set_origin(list)
}
