% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_clean.R
\name{sp_clean}
\alias{sp_clean}
\title{Clean species names}
\usage{
sp_clean(species)
}
\arguments{
\item{species}{vector of species}
}
\value{
Cleaned species vector
}
\description{
Clean species names and convert to short names used in Poland
}
\examples{
species_list = c('GLG ', 'Db', 'CZM P', 'Sosna Zwyczajna', 'Sesna', 'Abies alba', 'rowan')
sp_clean(species_list)
}
