% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_clean.R
\name{sp_clean}
\alias{sp_clean}
\title{Clean species}
\usage{
sp_clean(SPECIES_CD)
}
\arguments{
\item{SPECIES_CD}{vector of species}
}
\value{
Clean species vector
}
\description{
Clean species names
}
\examples{
species_list = c('GLG ', 'Db', 'CZM P', 'Sosna Zwyczajna', 'Sesna')
sp_clean(species_list)
}
