% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_group.R
\name{sp_group}
\alias{sp_group}
\title{Group species}
\usage{
sp_group(species, type = "GRP_V")
}
\arguments{
\item{species}{vector of species}

\item{type}{Grouping type (3 options): GRP_H (height group), GRP_V (volume group), GRP_P (growth group). Deafult: GRP_V description}
}
\value{
Species group name
}
\description{
Group species in eight groups (BK, OL, MD, BRZ, DB, ŚW, JD, SO),
necessary for the height and growth prediction
}
\examples{
species_list = c('GŁG', 'DB', 'CZM.P')
sp_group(species_list)
}
