% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_group.R
\name{sp_group}
\alias{sp_group}
\title{Group species}
\usage{
sp_group(species, type = "GRP_P")
}
\arguments{
\item{species}{vector of species}

\item{type}{Grouping type (5 options):
GRP_H (height group - Forest Data Bank based),
GRP_TH (top height group),
GRP_V (volume group),
GRP_P (growth group),
GRP_P_BDL (growth group - Forest Data Bank based).
Deafult: GRP_P}
}
\value{
Species group name
}
\description{
Group species in eight groups (BK, OL, MD, BRZ, DB, ŚW, JD, SO),
necessary for the height and growth prediction
}
\examples{
species_list = c('GŁG', 'DB', 'CZM.P')
sp_group(species_list, 'GRP_TH')
}
