% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_group.R
\name{sp_group}
\alias{sp_group}
\title{Group species}
\usage{
sp_group(species, type = "GRP_P")
}
\arguments{
\item{species}{vector of species}

\item{type}{grouping type (5 options: GRP_P, GRP_TH, GRP_V, GRP_H, GRP_P_BDL)}
}
\value{
Species group name
}
\description{
Group species in eight groups (BK, OL, MD, BRZ, DB, ŚW, JD, SO)
necessary for the height and growth prediction
}
\details{
\itemize{
\item GRP_P - (default) growth group
\item GRP_TH - top height group
\item GRP_V - volume group
\item GRP_H - height group; Forest Data Bank based
\item GRP_P_BDL - growth group; Forest Data Bank based
}
}
\examples{
species_list = c('AK', 'DB', 'CZM.P')
sp_group(species_list, 'GRP_TH')
}
