% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spg.R
\name{spg}
\alias{spg}
\title{Total coarse wood production}
\usage{
spg(age, height, species, region = "GLOB")
}
\arguments{
\item{age}{tree stand age}

\item{height}{tree stand top height}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB' (deafult GLOBAL)}
}
\value{
Total coarse wood production at input age in cubic meters per hectare
}
\description{
Predict total coarse wood production in the stand
based on the stand age, height and species. You cand calculate the production
based on the global and local parameters (forest regions).
}
\examples{
age = c(100, 101, 102)
height = c(32, 33, 34)
species = c('SO', 'DB', 'BK')
spg(age, height, species)
}
