% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_calc.R
\name{th_calc}
\alias{th_calc}
\title{Top Height (TH) calculation}
\usage{
th_calc(chm, f = "h_23")
}
\arguments{
\item{chm}{canopy height model (resolution equal or higher than 1 meter).
SpatRaster or RasterLayer}
}
\value{
SpatRaster of top height values (m) at 10x10 m resolution
}
\description{
Estimates the top height (TH) of a forest stand from a canopy height
model (CHM) raster. The function aggregates CHM pixels to a 10x10 m grid by
selecting the mean of the tallest cells within each grid cell, which approximates
the height of dominant trees. The aggregation factor is computed automatically
based on the resolution of the input raster, so any CHM with a resolution of
2 meters or finer is accepted.
}
\examples{
chm = terra::rast(system.file('raster/chm.tif', package = 'silvaR'))
terra::plot(chm)
th = th_calc(chm)
terra::plot(th)
}
