% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_calc.R
\name{th_calc}
\alias{th_calc}
\title{Top Height (TH) calculation}
\usage{
th_calc(chm, f = "h_23")
}
\arguments{
\item{chm}{canopy height model (resolution equal or higher than 1 meter).
SpatRaster or RasterLayer}

\item{f}{function used to calculate top height. Build-in functions: "h_23","h_sd"}
}
\value{
Top Height in the 10x10 m raster
}
\description{
Calculate TH of the tree stand in 10x10 m raster. Using two optional function
to obtain top height.
}
\details{
\itemize{
\item "h_23" - (default) mean of 1/3 highest cells to aggregate
\item "h_sd" - mean of cells higher than threshold: (percentile of 97 - 6.42)
}
}
\examples{
chm = terra::rast(system.file('raster/chm.tif', package = 'growthmodels'))
terra::plot(chm)
th = th_calc(chm, f = "h_sd")
terra::plot(th)
}
