% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_calc.R
\name{th_calc}
\alias{th_calc}
\title{Top Height (TH) calculation}
\usage{
th_calc(chm)
}
\arguments{
\item{chm}{canopy height model (resolution equal or lower than 1 m)}
}
\value{
Top Height in the 10x10 m raster
}
\description{
Calculate TH of the tree stand in 10x10 m raster. Using
1/3 highest cell values
}
\examples{
r =  terra::rast(ncol=100, nrow=100, nlyr=1, res = 0.5)
rv = runif(terra::ncell(r), min = 5, max = 36)
terra::values(r) = rv
terra::plot(r)
th = th_calc(r)
terra::plot(th)
}
