% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_growth.R
\name{v_growth}
\alias{v_growth}
\title{Stand volume growth prediction}
\usage{
v_growth(
  stand_id,
  years,
  age,
  height,
  volume,
  species,
  region = "GLOB",
  output_type = NULL
)
}
\arguments{
\item{stand_id}{stand id}

\item{years}{number of years that we want to predict the growth}

\item{age}{tree stand age}

\item{height}{Height at age T1 (average)}

\item{volume}{Volume at age T1}

\item{species}{Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{region}{Natural region - Abbreviation: 'I, 'II', 'III', 'IV', 'V', 'VI',
'VII', 'VIII', 'GLOB' (default)}

\item{output_type}{by default results will be returned as a vector, you can type 'df' to return data frame}
}
\value{
Numeric vector of cumulative volume growth over the prediction period (m3/ha),
or a data frame if output_type = 'df'
}
\description{
Predict stand volume increment over a specified number of years
using iterative year-by-year simulation. Each annual step calculates
height growth (th_growth()), site index, stocking index (stock_indx()),
total production (tvp()) and species share to derive the volume increment.
Growth is modified by region-specific parameters (ni1-ni4).
}
\examples{
years = 10
stand_id = c(1, 2, 3)
age = c(100, 101, 102)
height = c(32, 33, 34)
volume = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
region = c('I', 'II', 'GLOB')
v_growth(stand_id, years, age, height, volume, species, region, output_type = 'df')
}
\seealso{
\code{\link[=th_growth]{th_growth()}}, \code{\link[=tvp]{tvp()}}, \code{\link[=stock_indx]{stock_indx()}}
}
