% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_growth_climate.R
\name{v_growth_climate}
\alias{v_growth_climate}
\title{Volume prediction}
\usage{
v_growth_climate(
  stand_id,
  years,
  age,
  height,
  volume,
  species,
  rsum,
  maxtavg,
  output_type = NULL
)
}
\arguments{
\item{stand_id}{stand id}

\item{years}{number of years that we want to predict the growth}

\item{age}{tree stand age}

\item{height}{Height at age T1 (average)}

\item{volume}{Volume at age T1}

\item{species}{Species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}

\item{rsum}{precipitation sum #chodzi o sumę opadów z całego roku?}

\item{maxtavg}{maxium temperature #chodzi o maksymalną średnią temperaturę miesięczną?}

\item{output_type}{by default results will be returned as a vector, you can type 'df' to return data frame}
}
\value{
Growth sum in specific timeline
}
\description{
Predict tree volume growth in specific period of time based on H1,
and species with regard to climatic conditions
}
\examples{
years = 10
stand_id = c(1, 2, 3)
age = c(100, 101, 102)
height = c(32, 33, 34)
volume = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
rsum = c(800, 700, 600)
maxtavg = c(23, 17, 21.8)
v_growth_climate(stand_id, years, age, height, volume, species, rsum, maxtavg, output_type = 'df')
}
