% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tab.R
\name{v_tab}
\alias{v_tab}
\title{Tabular volume prediction}
\usage{
v_tab(age, height, species)
}
\arguments{
\item{age}{tree stand age (years)}

\item{height}{tree stand top height (m)}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}
}
\value{
Numeric vector of tabular volumes (m3/ha)
}
\description{
Predict stand volume based on height,
age and species using tabular reference data from
Szymkiewicz B. "Tablice zasobnosci i przyrostu drzewostanow".
Internally calculates Site Index (height at age 100) via th_growth() and
applies species-specific yield table parameters.
}
\examples{
age = c(100, 101, 102)
height = c(32, 33, 34)
species = c('SO', 'DB', 'BK')
v_tab(age, height, species)
}
\references{
Szymkiewicz T. (1952).
\emph{Tablice zasobności i przyrostu drzewostanów}.
Warszawa: Państwowe Wydawnictwo Rolnicze i Leśne.
}
