% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tab.R
\name{v_tab}
\alias{v_tab}
\title{Tabular volume prediction}
\usage{
v_tab(age, height, species)
}
\arguments{
\item{age}{tree stand age}

\item{height}{tree stand top height}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}
}
\value{
Tabular volume for species at input age in cubic meters per hectare
}
\description{
Predict stand volume based on height, age and species.
Function is using tabular reference data from
Szymkiewicz B. "Tablice zasobności i przyrostu drzewostanów"
}
\examples{
age = c(100, 101, 102)
height = c(32, 33, 34)
species = c('SO', 'DB', 'BK')
v_tab(age, height, species)
}
