% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tree.R
\name{v_tree}
\alias{v_tree}
\title{Single tree volume calculation}
\usage{
v_tree(dbh, height, species)
}
\arguments{
\item{dbh}{diameter at the brest height}

\item{height}{tree height}

\item{species}{tree species}
}
\value{
Tree volume
}
\description{
Calculate single tree volume
}
\examples{
dbh = c(32, 33, 34)
height = c(27, 28, 29)
species = c('SO', 'DB', 'OL SZ')
v_tree(dbh, height, species)
}
