% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tree_bruchwald.R
\name{v_tree_bruchwald}
\alias{v_tree_bruchwald}
\title{Single tree volume calculation (Bruchwald method)}
\usage{
v_tree_bruchwald(
  plot_id,
  tree_id,
  species,
  age,
  layer,
  dbh,
  height,
  origin = NA
)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{tree_id}{Unique tree id at the plot}

\item{species}{Tree species}

\item{age}{Tree age (years)}

\item{layer}{Stand structure layer}

\item{dbh}{Tree DBH (cm)}

\item{height}{Tree height (m)}

\item{origin}{L (lowland) or M (mountain). This parameter needs to be provided if the species is BK}
}
\value{
Numeric vector of tree stem volumes (m3)
}
\description{
Calculate individual tree stem volume using species-specific
Bruchwald (2000) equations with form factors and bark coefficients.
Unlike v_tree(), this method uses plot-level auxiliary variables
(RMS DBH, Lorey's height) and applies distinct equation sets depending on
species group: form-factor equations for SO, SW, JD, DB, BK, BRZ, OL, DG, and
linear/power-law equations for MD, OS, GB, TP, LP, CZR.
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
tree_id = c(1, 2 ,3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
age = c(40, 40, 40, 40, 60, 45, 50, 50, 50, 60, 60, 60, 60, 60)
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
dbh = c(33, 32, 31, NA, 28, 47, 12, 22, 41, 48, 27, 42, 25, 33)
origin = c(NA, NA, NA, NA, 'M', NA, NA, NA, NA, NA, NA, NA, NA, NA)
v_tree_bruchwald(plot_id, tree_id, species, age, layer, dbh, height, origin)
}
\references{
Bruchwald et all. (2000).
\emph{Wzory empiryczne do okreslania wysokosci i piersnicowej liczby ksztaltu grubizny drzewa.}.
Sylwan, 144, 5–12
}
