% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tree_bruchwald.R
\name{v_tree_bruchwald}
\alias{v_tree_bruchwald}
\title{Single tree volume calculation}
\usage{
v_tree_bruchwald(
  plot_id,
  tree_id,
  species,
  age,
  layer,
  dbh,
  height,
  origin = NA
)
}
\arguments{
\item{plot_id}{Unique plot id}

\item{tree_id}{Unique tree id at the plot}

\item{species}{Tree species}

\item{age}{Tree age}

\item{layer}{Stand structure layer}

\item{dbh}{Tree dbh}

\item{height}{Tree heighs}

\item{origin}{L (lowland) or M (mountain). This parameter needs to be provided if the species is BK}
}
\value{
Tree volume
}
\description{
Calculate single tree volume
}
\examples{
plot_id = c(1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2)
tree_id = c(1, 2 ,3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14)
species = c('SO', 'SO', 'DB', 'JW', 'BK', 'DB', 'DB', 'ŚW', 'ŚW', 'ŚW', 'SO', 'SO', 'SO', 'SO')
age = c(40, 40, 40, 40, 60, 45, 50, 50, 50, 60, 60, 60, 60, 60)
layer = c(1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 1, 1)
height = c(21, 13, 24, NA, 12, NA, 18, NA, NA, 31, 32, 24, 25, NA)
dbh = c(33, 32, 31, NA, 28, 47, 12, 22, 41, 48, 27, 42, 25, 33)
origin = c(NA, NA, NA, NA, 'M', NA, NA, NA, NA, NA, NA, NA, NA, NA)
v_tree_bruchwald(plot_id, tree_id, species, age, layer, dbh, height, origin)
}
