% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tree_bruchwald.R
\name{v_tree_bruchwald}
\alias{v_tree_bruchwald}
\title{Single tree volume calculation}
\usage{
v_tree_bruchwald(
  dbh,
  height,
  av_H,
  av_DBH,
  species,
  origin = NA,
  bark_param = T
)
}
\arguments{
\item{dbh}{diameter at the brest height}

\item{height}{tree height}

\item{av_H}{average height}

\item{av_DBH}{average DBH}

\item{species}{tree species}

\item{origin}{L (lowland) or M (mountain). This parameter needs to be provided if the species is BK}

\item{bark_param}{Consideration of the bark parameter (logical)}
}
\value{
Tree volume
}
\description{
Calculate single tree volume
}
\examples{
dbh = c(31, 31, 31, NA, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31)
height = c(27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27, 27)
av_H = c(26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26, 26)
av_DBH = c(31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5, 31.5)
species = c('SO', 'JD', 'DB', 'BK', 'BRZ', 'OL', 'DG', 'MD', 'OS', 'GB', 'TP', 'LP', 'CZR', 'OL CZ')
origin = c(NA, NA, NA, 'M', NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
}
