% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zd_share.R
\name{zd_share}
\alias{zd_share}
\title{Stocking index}
\usage{
zd_share(stand_id, volume, age, height, species)
}
\arguments{
\item{stand_id}{stand id}

\item{volume}{volume in the stand by age-species group}

\item{age}{tree stand age by age-species group}

\item{height}{tree stand height by age-species group}

\item{species}{species - Abbreviation: 'BK', 'OL', 'MD', 'BRZ', 'DB', 'ŚW',
'JD', 'SO'}
}
\value{
Stock index as a ratio of actual stock to tabular volume
}
\description{
Calculation of the stocking index of the stand on the basis of
the calculated volume and the tabular volume (see v_tab() function) by age-species group.
}
\examples{
stand_id = c(1, 2, 3)
age = c(100, 101, 102)
height = c(32, 33, 34)
volume = c(150, 160, 170)
species = c('SO', 'DB', 'BK')
}
